Refactoring_model
モデルをDRY原則に則りリファクタリングした
