+++ app/Answer.php
@@ -6,6 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Answer extends Model
 {
+    use VotableTrait;
+
     protected $fillable = ['body', 'user_id'];
 
     public function question ()
@@ -54,20 +56,5 @@ class Answer extends Model
     public function IsBest()
     {
         return $this->id === $this->question->best_answer_id;
-    }
-
-    public function votes()
-    {
-        return $this->morphToMany(User::class, 'votable');
-    }
-
-    public function upVotes()
-    {
-        return $this->votes()->wherePivot('vote', 1);
-    }
-
-    public function downVotes()
-    {
-        return $this->votes()->wherePivot('vote', -1);
     }
 }
