+++ app/User.php
@@ -78,36 +78,31 @@ class User extends Authenticatable
     public function voteQuestion(Question $question, $vote)
     {
         $voteQuestions = $this->voteQuestions();
-        if ($voteQuestions->where('votable_id', $question->id)->exists()) {
-            $voteQuestions->updateExistingPivot($question, ['vote' => $vote]);
-        }
-        else {
-            $voteQuestions->attach($question, ['vote' => $vote]);
-        }
 
-        $question->load('votes');
-        $upVotes = (int) $question->upVotes()->sum('vote');
-        $downVotes = (int) $question->downVotes()->sum('vote');
-
-        $question->votes_count = $upVotes + $downVotes;
-        $question->save();
+        $this->_vote($voteQuestions, $question, $vote);
     }
 
     public function voteAnswer(Answer $answer, $vote)
     {
         $voteAnswers = $this->voteAnswers();
-        if ($voteAnswers->where('votable_id', $answer->id)->exists()) {
-            $voteAnswers->updateExistingPivot($answer, ['vote' => $vote]);
+
+        $this->_vote($voteAnswers, $answer, $vote);
+    }
+
+    private function _vote($relationship, $model, $vote)
+    {
+        if ($relationship->where('votable_id', $model->id)->exists()) {
+            $relationship->updateExistingPivot($model, ['vote' => $vote]);
         }
         else {
-            $voteAnswers->attach($answer, ['vote' => $vote]);
+            $relationship->attach($model, ['vote' => $vote]);
         }
 
-        $answer->load('votes');
-        $upVotes = (int) $answer->upVotes()->sum('vote');
-        $downVotes = (int) $answer->downVotes()->sum('vote');
+        $model->load('votes');
+        $upVotes = (int) $model->upVotes()->sum('vote');
+        $downVotes = (int) $model->downVotes()->sum('vote');
 
-        $answer->votes_count = $upVotes + $downVotes;
-        $answer->save();
+        $model->votes_count = $upVotes + $downVotes;
+        $model->save();
     }
 }
