+++ app/votableTrait.php
@@ -1,0 +1,20 @@
+<?php
+namespace App;
+
+trait VotableTrait
+{
+  public function votes()
+  {
+      return $this->morphToMany(User::class, 'votable');
+  }
+
+  public function upVotes()
+  {
+      return $this->votes()->wherePivot('vote', 1);
+  }
+
+  public function downVotes()
+  {
+      return $this->votes()->wherePivot('vote', -1);
+  }
+}
